+++ R/AllGenerics.R
@@ -280,9 +280,10 @@ setGeneric("elmt", function(object,
 #'   hypothesis, with as many entries as the rows in `lhs`. Defaults to `NULL`.
 #'   See ‘Details’.
 #' @param lhs A numeric matrix or a vector (treated as a row matrix) for the
-#'   left-hand side of hypothesis. Each row gives a linear combination of the
+#'   left-hand side of a hypothesis. Each row gives a linear combination of the
 #'   parameters in `object`. The number of columns should be equal to the number
-#'   of parameters. Defaults to `NULL`. See ‘Details’.
+#'   of parameters. Or a character vector with a symbolic description of the
+#'   hypothesis is allowed. Defaults to `NULL`. See ‘Details’.
 #' @param alpha A single numeric for the significance level. Defaults to `0.05`.
 #' @param calibrate A single character for the calibration method. It is
 #'   case-insensitive and must be one of `"chisq"`, `"boot"`, or `"f"`.
@@ -327,10 +328,9 @@ setGeneric("elmt", function(object,
 #' @usage NULL
 #' @examples
 #' ## F calibration for the mean
-#' set.seed(533414)
-#' x <- rnorm(100)
-#' fit <- el_mean(x, 0)
-#' elt(fit, rhs = 0.3, calibrate = "f")
+#' data("precip")
+#' fit <- el_mean(precip, 32)
+#' elt(fit, rhs = 32, calibrate = "f")
 #'
 #' ## Test of no treatment effect
 #' data("clothianidin")
@@ -341,6 +341,11 @@ setGeneric("elmt", function(object,
 #' ), byrow = TRUE, nrow = 3)
 #' fit2 <- el_lm(clo ~ -1 + trt, clothianidin)
 #' elt(fit2, lhs = contrast)
+#'
+#' ## A symbolic description of the same hypothesis
+#' elt(fit2, lhs = c(
+#'   "trtNaked - trtFungicide", "trtFungicide - trtLow", "trtLow - trtHigh"
+#' ))
 #' @exportMethod elt
 #' @srrstats {G5.1} `clothianidin` data set is exported.
 setGeneric("elt", function(object,
@@ -595,6 +600,16 @@ setGeneric("weights", function(object, ...) standardGeneric("weights"))
 setGeneric("getData", function(x) standardGeneric("getData"))
 setMethod("getData", "EL", function(x) {
   x@data
+})
+
+
+setGeneric("getEstimates", function(x) standardGeneric("getEstimates"))
+setMethod("getEstimates", "EL", function(x) {
+  x@coefficients
+})
+
+setMethod("getEstimates", "QGLM", function(x) {
+  c(x@coefficients, x@dispersion)
 })